+++ lib/src/widgets/docking_item_widget.dart
@@ -43,7 +43,7 @@ class DockingItemWidget extends DraggableWidget {
       buttons = [];
       buttons.addAll(item.buttons!);
     }
-    bool maximizable =
+    final bool maximizable =
         item.maximizable != null ? item.maximizable! : this.maximizable;
     if (maximizable) {
       if (buttons == null) {